@@ -7,17 +7,16 @@ const REGISTER = gql`
   mutation register($registration: Registration!) {
     register(input: $registration) {
       res
-      error
     }
   }
 `;
 
 @Injectable()
 export class RegisterService {
-  constructor(private readonly apollo: Apollo) {}
+  constructor(private readonly _apollo: Apollo) {}
 
   register(r: RegistrationReq) {
-    return this.apollo.mutate({
+    return this._apollo.mutate({
       mutation: REGISTER,
       variables: {
         registration: r,
